Add vector of int64 in Tensor
- change interface to support and check DType when populating
- rename interface for getData to getDoubleData
- add functions to access IntData: getIntData
- update registery to allow constant op to receive 2 types of data
- update code calling getData to check for isDoubleType() and calling
the appropriate data

+++ src/pass/finalization/sparsity.cpp
@@ -141,7 +141,7 @@ void setSparsityFnc(const mv::pass::PassEntry& pass, mv::ComputationModel& model
 
             //Create Tensor with Sparsity Data
             mv::Shape sparsityShape(ndims);
-            std::vector<double> data(sparsityShape.totalSize(), 0);
+            std::vector<int64_t> data(sparsityShape.totalSize(), 0);
             //mv::Tensor sparsityTensor("backup", sparsityShape, mv::DType("UInt8"), mv::Order("WHCN"), data);
             auto sparsityTensor = dm.defineTensor(opIterator->getName() + "_sparse_dw", sparsityShape, mv::DType("UInt8"), mv::Order("NCHW"), data);
